requirements.txt
opencv-python
tensorflow
keras
numpy
pillow
scikit-learn
scrapy